+++ sdk/identity/src/token_credentials/azure_cli_credentials.rs
