+++ sdk/identity/src/token_credentials/default_credentials.rs
