+++ sdk/identity/src/token_credentials/environment_credentials.rs
