+++ sdk/identity/src/token_credentials/imds_managed_identity_credentials.rs
